switch to library with binary
